Move Mastodon theme handling to custom vite plugin
