+++ config/vite/plugin-mastodon-themes.ts
@@ -1,0 +1,52 @@
+/* This plugins handles Mastodon's theme system
+ */
+
+import fs from 'node:fs/promises';
+import path from 'node:path';
+
+import yaml from 'js-yaml';
+import type { Plugin } from 'vite';
+
+export function MastodonThemes(projectRoot: string): Plugin {
+  return {
+    name: 'mastodon-themes',
+    config: async () => {
+      const jsRoot = path.resolve(projectRoot, 'app/javascript');
+      const themesFile = path.resolve(projectRoot, 'config/themes.yml');
+      const entrypoints: Record<string, string> = {};
+
+      // Get all files mentioned in the themes.yml file.
+      const themesString = await fs.readFile(themesFile, 'utf8');
+      const themes = yaml.load(themesString, {
+        filename: 'themes.yml',
+        schema: yaml.FAILSAFE_SCHEMA,
+      });
+
+      if (!themes || typeof themes !== 'object') {
+        throw new Error('Invalid themes.yml file');
+      }
+
+      for (const themePath of Object.values(themes)) {
+        if (
+          typeof themePath !== 'string' ||
+          themePath.split('.').length !== 2 || // Ensure it has exactly one period
+          !themePath.endsWith('css')
+        ) {
+          console.warn(
+            `Invalid theme path "${themePath}" in themes.yml, skipping`,
+          );
+          continue;
+        }
+        entrypoints[path.basename(themePath)] = path.resolve(jsRoot, themePath);
+      }
+
+      return {
+        build: {
+          rollupOptions: {
+            input: entrypoints,
+          },
+        },
+      };
+    },
+  };
+}
